refactor: remove MediaController
